+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Lockable and EventHandler to be non-copyable.
+- Mutex to abort when locked or unlocked from an ISR.
 
 ### Fixed
 - StreamServer task not being stopped before the derived object is destroyed.